+++ InvoiceManagementStudio/Core/Definition/Entity/Receiver/IEntityReceiverDefinition.cs
@@ -1,4 +1,4 @@
-namespace InvoiceManagementStudio.Core.Definition.Entity
+namespace InvoiceManagementStudio.Core.Definition.Entity.Receiver
 {
 
     public interface IEntityReceiverDefinition : IEntityDefinition
