+++ InvoiceManagementStudio/Core/Definition/Invoice/IInvoiceDefinition.cs
@@ -2,6 +2,10 @@ using System;
 using System.Collections.Generic;
 
 using InvoiceManagementStudio.Core.Definition.Entity;
+using InvoiceManagementStudio.Core.Definition.Entity.Receiver;
+using InvoiceManagementStudio.Core.Definition.Entity.Supplier;
+using InvoiceManagementStudio.Core.Definition.Item;
+using InvoiceManagementStudio.Core.Definition.Payment;
 
 
 namespace InvoiceManagementStudio.Core.Definition.Invoice
