+++ InvoiceManagementStudio/Core/Definition/Invoice/IInvoiceRepository.cs
@@ -1,0 +1,9 @@
+namespace InvoiceManagementStudio.Core.Definition.Invoice
+{
+
+    public interface IInvoiceRepository : ICrudRepository<ulong, IInvoiceDefinition>
+    {
+        
+    }
+
+}
