+++ InvoiceManagementStudio/Core/Definition/Invoice/IInvoiceService.cs
@@ -1,7 +1,7 @@
 namespace InvoiceManagementStudio.Core.Definition.Invoice
 {
 
-    public interface IInvoicePaymentDefinition
+    public interface IInvoiceService  : ICrudService<ulong, IInvoiceDefinition>
     {
         
     }
