+++ InvoiceManagementStudio/Core/Definition/Item/EUnitType.cs
@@ -1,4 +1,4 @@
-namespace InvoiceManagementStudio.Core.Definition.Invoice
+namespace InvoiceManagementStudio.Core.Definition.Item
 {
 
     /// <summary>
