+++ InvoiceManagementStudio/Core/Definition/Item/IInvoiceItemRepository.cs
@@ -1,0 +1,9 @@
+namespace InvoiceManagementStudio.Core.Definition.Item
+{
+
+    public interface IInvoiceItemRepository : ICrudRepository<ulong, IInvoiceItemDefinition>
+    {
+        
+    }
+
+}
