+++ InvoiceManagementStudio/Core/Definition/Item/IInvoiceItemService.cs
@@ -1,0 +1,9 @@
+namespace InvoiceManagementStudio.Core.Definition.Item
+{
+
+    public interface IInvoiceItemService  : ICrudService<ulong, IInvoiceItemDefinition>
+    {
+        
+    }
+
+}
