+++ InvoiceManagementStudio/Core/Definition/Payment/IInvoicePaymentDefinition.cs
@@ -1,0 +1,9 @@
+namespace InvoiceManagementStudio.Core.Definition.Payment
+{
+
+    public interface IInvoicePaymentDefinition
+    {
+        
+    }
+
+}
