+++ InvoiceManagementStudio/Core/Definition/Payment/IInvoicePaymentService.cs
@@ -1,0 +1,9 @@
+namespace InvoiceManagementStudio.Core.Definition.Payment
+{
+
+    public interface IInvoicePaymentService : ICrudService<ulong, IInvoicePaymentDefinition>
+    {
+        
+    }
+
+}
