--- InvoiceManagementStudio/Core/Definition/Repository/Invoice/IInvoiceRepository.cs
@@ -1,12 +1,0 @@
-using InvoiceManagementStudio.Core.Definition.Invoice;
-
-
-namespace InvoiceManagementStudio.Core.Definition.Repository.Invoice
-{
-
-    public interface IInvoiceRepository : ICrudRepository<ulong, IInvoiceDefinition>
-    {
-        
-    }
-
-}
