--- InvoiceManagementStudio/Core/Definition/Repository/Item/IInvoiceItemRepository.cs
@@ -1,12 +1,0 @@
-using InvoiceManagementStudio.Core.Definition.Invoice;
-
-
-namespace InvoiceManagementStudio.Core.Definition.Repository.Item
-{
-
-    public interface IInvoiceItemRepository : ICrudRepository<ulong, IInvoiceItemDefinition>
-    {
-        
-    }
-
-}
