--- InvoiceManagementStudio/Core/Definition/Service/Invoice/IInvoiceService.cs
@@ -1,12 +1,0 @@
-using InvoiceManagementStudio.Core.Definition.Invoice;
-
-
-namespace InvoiceManagementStudio.Core.Definition.Service.Invoice
-{
-
-    public interface IInvoiceService  : ICrudService<ulong, IInvoiceDefinition>
-    {
-        
-    }
-
-}
