--- InvoiceManagementStudio/Core/Definition/Service/Item/IInvoiceItemService.cs
@@ -1,12 +1,0 @@
-using InvoiceManagementStudio.Core.Definition.Invoice;
-
-
-namespace InvoiceManagementStudio.Core.Definition.Service.Item
-{
-
-    public interface IInvoiceItemService  : ICrudService<ulong, IInvoiceItemDefinition>
-    {
-        
-    }
-
-}
