--- InvoiceManagementStudio/Core/Definition/Service/Payment/IInvoicePaymentService.cs
@@ -1,12 +1,0 @@
-using InvoiceManagementStudio.Core.Definition.Invoice;
-
-
-namespace InvoiceManagementStudio.Core.Definition.Service.Payment
-{
-
-    public interface IInvoicePaymentService : ICrudService<ulong, IInvoicePaymentDefinition>
-    {
-        
-    }
-
-}
